pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vpn-service"
version = "0.1.0"
description = "Commercial VPN service on FastAPI with Remnawave integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.0",
  "asyncpg>=0.30.0",
  "celery[redis]>=5.4.0",
  "email-validator>=2.2.0",
  "fastapi>=0.115.0",
  "greenlet>=3.1.1",
  "httpx>=0.28.0",
  "passlib[bcrypt]>=1.7.4",
  "pydantic-settings>=2.6.1",
  "pyjwt>=2.10.1",
  "python-multipart>=0.0.17",
  "redis>=5.2.1",
  "remnawave>=2.6.3",
  "slowapi>=0.1.9",
  "sqlalchemy>=2.0.36",
  "structlog>=24.4.0",
  "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
  "aiosqlite>=0.20.0",
  "faker>=33.1.0",
  "pytest>=8.3.4",
  "pytest-asyncio>=0.25.0",
  "respx>=0.22.0",
]

[tool.setuptools]
packages = ["src"]

[tool.setuptools.package-dir]
"" = "."

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"
